Add TokenScope to store tokens with their scope for users, to allow people to choose if they want none, any of them, etc
